+++ src/System.Windows.Forms/src/System/Windows/Forms/PropertyGridInternal/PropertyGridView.cs
@@ -6031,823 +6031,6 @@ namespace System.Windows.Forms.PropertyGridInternal
             return;
         }
 
-        internal class GridViewListBox : ListBox
-        {
-            internal bool fInSetSelectedIndex;
-            private readonly PropertyGridView _owningPropertyGridView;
-
-            public GridViewListBox(PropertyGridView gridView)
-            {
-                if (gridView == null)
-                {
-                    throw new ArgumentNullException(nameof(gridView));
-                }
-
-                base.IntegralHeight = false;
-                _owningPropertyGridView = gridView;
-                base.BackColor = gridView.BackColor;
-            }
-
-            protected override CreateParams CreateParams
-            {
-                get
-                {
-                    CreateParams cp = base.CreateParams;
-                    cp.Style &= ~(int)User32.WS.BORDER;
-                    cp.ExStyle &= ~(int)User32.WS_EX.CLIENTEDGE;
-                    return cp;
-                }
-            }
-
-            /// <summary>
-            ///  Gets the owning PropertyGridView.
-            /// </summary>
-            internal PropertyGridView OwningPropertyGridView
-            {
-                get
-                {
-                    return _owningPropertyGridView;
-                }
-            }
-
-            /// <summary>
-            ///  Indicates whether or not the control supports UIA Providers via
-            ///  IRawElementProviderFragment/IRawElementProviderFragmentRoot interfaces
-            /// </summary>
-            internal override bool SupportsUiaProviders => true;
-
-            /// <summary>
-            ///  Constructs the new instance of the accessibility object for this control.
-            /// </summary>
-            /// <returns>The accessibility object instance.</returns>
-            protected override AccessibleObject CreateAccessibilityInstance()
-            {
-                return new GridViewListBoxAccessibleObject(this);
-            }
-
-            public virtual bool InSetSelectedIndex()
-            {
-                return fInSetSelectedIndex;
-            }
-
-            protected override void OnSelectedIndexChanged(EventArgs e)
-            {
-                fInSetSelectedIndex = true;
-                base.OnSelectedIndexChanged(e);
-                fInSetSelectedIndex = false;
-            }
-        }
-
-        private class GridViewListBoxItemAccessibleObject : ListBox.ListBoxItemAccessibleObject
-        {
-            private readonly GridViewListBox _owningGridViewListBox;
-            private readonly object _owningItem;
-
-            public GridViewListBoxItemAccessibleObject(GridViewListBox owningGridViewListBox, object owningItem)
-                : base(owningGridViewListBox, owningItem, owningGridViewListBox.AccessibilityObject as ListBox.ListBoxAccessibleObject)
-            {
-                _owningGridViewListBox = owningGridViewListBox;
-                _owningItem = owningItem;
-            }
-
-            internal override UiaCore.IRawElementProviderFragmentRoot FragmentRoot
-            {
-                get
-                {
-                    return _owningGridViewListBox.AccessibilityObject;
-                }
-            }
-
-            internal override object GetPropertyValue(UiaCore.UIA propertyID)
-            {
-                switch (propertyID)
-                {
-                    case UiaCore.UIA.AccessKeyPropertyId:
-                        return KeyboardShortcut;
-                    default:
-                        return base.GetPropertyValue(propertyID);
-                }
-            }
-
-            /// <summary>
-            ///  Indicates whether specified pattern is supported.
-            /// </summary>
-            /// <param name="patternId">The pattern ID.</param>
-            /// <returns>True if specified </returns>
-            internal override bool IsPatternSupported(UiaCore.UIA patternId)
-            {
-                if (patternId == UiaCore.UIA.InvokePatternId)
-                {
-                    return true;
-                }
-
-                return base.IsPatternSupported(patternId);
-            }
-
-            /// <summary>
-            ///  Gets or sets the accessible name.
-            /// </summary>
-            public override string Name
-            {
-                get
-                {
-                    if (_owningGridViewListBox != null)
-                    {
-                        return _owningItem.ToString();
-                    }
-
-                    return base.Name;
-                }
-                set => base.Name = value;
-            }
-
-            /// <summary>
-            ///  Gets the runtime ID.
-            /// </summary>
-            internal override int[] RuntimeId
-            {
-                get
-                {
-                    var runtimeId = new int[3];
-                    runtimeId[0] = RuntimeIDFirstItem;
-                    runtimeId[1] = (int)(long)_owningGridViewListBox.Handle;
-                    runtimeId[2] = _owningItem.GetHashCode();
-
-                    return runtimeId;
-                }
-            }
-        }
-
-        /// <summary>
-        ///  Represents the PropertyGridView ListBox accessibility object.
-        /// </summary>
-        private class GridViewListBoxAccessibleObject : ListBox.ListBoxAccessibleObject
-        {
-            private readonly PropertyGridView _owningPropertyGridView;
-
-            /// <summary>
-            ///  Constructs the new instance of GridViewListBoxAccessibleObject.
-            /// </summary>
-            /// <param name="owningGridViewListBox">The owning GridViewListBox.</param>
-            public GridViewListBoxAccessibleObject(GridViewListBox owningGridViewListBox) : base(owningGridViewListBox)
-            {
-                if (owningGridViewListBox.OwningPropertyGridView == null)
-                {
-                    throw new ArgumentNullException(nameof(owningGridViewListBox.OwningPropertyGridView));
-                }
-
-                _owningPropertyGridView = owningGridViewListBox.OwningPropertyGridView;
-            }
-
-            /// <summary>
-            ///  Request to return the element in the specified direction.
-            /// </summary>
-            /// <param name="direction">Indicates the direction in which to navigate.</param>
-            /// <returns>Returns the element in the specified direction.</returns>
-            internal override UiaCore.IRawElementProviderFragment FragmentNavigate(UiaCore.NavigateDirection direction)
-            {
-                if (direction == UiaCore.NavigateDirection.Parent && _owningPropertyGridView.SelectedGridEntry != null)
-                {
-                    return _owningPropertyGridView.SelectedGridEntry.AccessibilityObject;
-                }
-                else if (direction == UiaCore.NavigateDirection.NextSibling)
-                {
-                    return _owningPropertyGridView.Edit.AccessibilityObject;
-                }
-
-                return base.FragmentNavigate(direction);
-            }
-
-            public override string Name
-            {
-                get => base.Name ?? SR.PropertyGridEntryValuesListDefaultAccessibleName;
-                set => base.Name = value;
-            }
-
-            /// <summary>
-            ///  Return the element that is the root node of this fragment of UI.
-            /// </summary>
-            internal override UiaCore.IRawElementProviderFragmentRoot FragmentRoot
-            {
-                get
-                {
-                    return _owningPropertyGridView.AccessibilityObject;
-                }
-            }
-        }
-
-        private class GridViewEdit : TextBox, IMouseHookClient
-        {
-            internal bool fInSetText;
-            internal bool filter;
-            internal PropertyGridView psheet;
-            private bool dontFocusMe;
-            private int lastMove;
-
-            private readonly MouseHook mouseHook;
-
-            // We do this because the Focus call above doesn't always stick, so
-            // we make the Edit think that it doesn't have focus.  this prevents
-            // ActiveControl code on the containercontrol from moving focus elsewhere
-            // when the dropdown closes.
-            public bool DontFocus
-            {
-                set
-                {
-                    dontFocusMe = value;
-                }
-            }
-
-            public virtual bool Filter
-            {
-                get { return filter; }
-
-                set
-                {
-                    filter = value;
-                }
-            }
-
-            /// <summary>
-            ///  Indicates whether or not the control supports UIA Providers via
-            ///  IRawElementProviderFragment/IRawElementProviderFragmentRoot interfaces
-            /// </summary>
-            internal override bool SupportsUiaProviders => true;
-
-            public override bool Focused
-            {
-                get
-                {
-                    if (dontFocusMe)
-                    {
-                        return false;
-                    }
-                    return base.Focused;
-                }
-            }
-
-            public override string Text
-            {
-                get => base.Text;
-                set
-                {
-                    fInSetText = true;
-                    base.Text = value;
-                    fInSetText = false;
-                }
-            }
-
-            public bool DisableMouseHook
-            {
-                set
-                {
-                    mouseHook.DisableMouseHook = value;
-                }
-            }
-
-            public virtual bool HookMouseDown
-            {
-                get
-                {
-                    return mouseHook.HookMouseDown;
-                }
-                set
-                {
-                    mouseHook.HookMouseDown = value;
-                    if (value)
-                    {
-                        Focus();
-                    }
-                }
-            }
-
-            public GridViewEdit(PropertyGridView psheet)
-            {
-                this.psheet = psheet;
-                mouseHook = new MouseHook(this, this, psheet);
-            }
-
-            /// <summary>
-            ///  Creates a new AccessibleObject for this GridViewEdit instance.
-            ///  The AccessibleObject instance returned by this method overrides several UIA properties.
-            /// </summary>
-            /// <returns>
-            ///  AccessibleObject for this GridViewEdit instance.
-            /// </returns>
-            protected override AccessibleObject CreateAccessibilityInstance()
-            {
-                return new GridViewEditAccessibleObject(this);
-            }
-
-            protected override void DestroyHandle()
-            {
-                mouseHook.HookMouseDown = false;
-                base.DestroyHandle();
-            }
-
-            protected override void Dispose(bool disposing)
-            {
-                if (disposing)
-                {
-                    mouseHook.Dispose();
-                }
-                base.Dispose(disposing);
-            }
-
-            public void FilterKeyPress(char keyChar)
-            {
-                if (IsInputChar(keyChar))
-                {
-                    Focus();
-                    SelectAll();
-                    User32.PostMessageW(this, User32.WM.CHAR, (IntPtr)keyChar);
-                }
-            }
-
-            /// <summary>
-            ///  Overridden to handle TAB key.
-            /// </summary>
-            protected override bool IsInputKey(Keys keyData)
-            {
-                switch (keyData & Keys.KeyCode)
-                {
-                    case Keys.Escape:
-                    case Keys.Tab:
-                    case Keys.F4:
-                    case Keys.F1:
-                    case Keys.Return:
-                        return false;
-                }
-                if (psheet.NeedsCommit)
-                {
-                    return false;
-                }
-                return base.IsInputKey(keyData);
-            }
-
-            /// <summary>
-            ///  Overridden to handle TAB key.
-            /// </summary>
-            protected override bool IsInputChar(char keyChar)
-            {
-                switch ((Keys)(int)keyChar)
-                {
-                    case Keys.Tab:
-                    case Keys.Return:
-                        return false;
-                }
-                return base.IsInputChar(keyChar);
-            }
-
-            protected override void OnGotFocus(EventArgs e)
-            {
-                base.OnGotFocus(e);
-
-                this.AccessibilityObject.RaiseAutomationEvent(UiaCore.UIA.AutomationFocusChangedEventId);
-            }
-
-            protected override void OnKeyDown(KeyEventArgs ke)
-            {
-                // this is because on a dialog we may
-                // not get a chance to pre-process
-                //
-                if (ProcessDialogKey(ke.KeyData))
-                {
-                    ke.Handled = true;
-                    return;
-                }
-
-                base.OnKeyDown(ke);
-            }
-
-            protected override void OnKeyPress(KeyPressEventArgs ke)
-            {
-                if (!IsInputChar(ke.KeyChar))
-                {
-                    ke.Handled = true;
-                    return;
-                }
-                base.OnKeyPress(ke);
-            }
-
-            public bool OnClickHooked()
-            {
-                // can we commit this value?
-                // eat the value if we failed to commit.
-                return !psheet._Commit();
-            }
-
-            protected override void OnMouseEnter(EventArgs e)
-            {
-                base.OnMouseEnter(e);
-
-                if (!Focused)
-                {
-                    Graphics g = CreateGraphics();
-                    if (psheet.SelectedGridEntry != null &&
-                        ClientRectangle.Width <= psheet.SelectedGridEntry.GetValueTextWidth(Text, g, Font))
-                    {
-                        psheet.ToolTip.ToolTip = PasswordProtect ? "" : Text;
-                    }
-                    g.Dispose();
-                }
-            }
-
-            protected override bool ProcessCmdKey(ref Message msg, Keys keyData)
-            {
-                // make sure we allow the Edit to handle ctrl-z
-                switch (keyData & Keys.KeyCode)
-                {
-                    case Keys.Z:
-                    case Keys.C:
-                    case Keys.X:
-                    case Keys.V:
-                        if (
-                           ((keyData & Keys.Control) != 0) &&
-                           ((keyData & Keys.Shift) == 0) &&
-                           ((keyData & Keys.Alt) == 0))
-                        {
-                            return false;
-                        }
-                        break;
-
-                    case Keys.A:
-                        if (
-                           ((keyData & Keys.Control) != 0) &&
-                           ((keyData & Keys.Shift) == 0) &&
-                           ((keyData & Keys.Alt) == 0))
-                        {
-                            SelectAll();
-                            return true;
-                        }
-
-                        break;
-
-                    case Keys.Insert:
-                        if (((keyData & Keys.Alt) == 0))
-                        {
-                            if (((keyData & Keys.Control) != 0) ^ ((keyData & Keys.Shift) == 0))
-                            {
-                                return false;
-                            }
-                        }
-                        break;
-
-                    case Keys.Delete:
-                        if (
-                          ((keyData & Keys.Control) == 0) &&
-                          ((keyData & Keys.Shift) != 0) &&
-                          ((keyData & Keys.Alt) == 0))
-                        {
-                            return false;
-                        }
-                        else if (
-                              ((keyData & Keys.Control) == 0) &&
-                              ((keyData & Keys.Shift) == 0) &&
-                              ((keyData & Keys.Alt) == 0)
-                                )
-                        {
-                            // if this is just the delete key and we're on a non-text editable property that is resettable,
-                            // reset it now.
-                            //
-                            if (psheet.SelectedGridEntry != null && !psheet.SelectedGridEntry.Enumerable && !psheet.SelectedGridEntry.IsTextEditable && psheet.SelectedGridEntry.CanResetPropertyValue())
-                            {
-                                object oldValue = psheet.SelectedGridEntry.PropertyValue;
-                                psheet.SelectedGridEntry.ResetPropertyValue();
-                                psheet.UnfocusSelection();
-                                psheet.OwnerGrid.OnPropertyValueSet(psheet.SelectedGridEntry, oldValue);
-                            }
-                        }
-                        break;
-                }
-                return base.ProcessCmdKey(ref msg, keyData);
-            }
-
-            /// <summary>
-            ///  Overrides Control.ProcessDialogKey to handle the Escape and Return
-            ///  keys.
-            /// </summary>
-            protected override bool ProcessDialogKey(Keys keyData)
-            {
-                // We don't do anything with modified keys here.
-                //
-                if ((keyData & (Keys.Shift | Keys.Control | Keys.Alt)) == 0)
-                {
-                    switch (keyData & Keys.KeyCode)
-                    {
-                        case Keys.Return:
-                            bool fwdReturn = !psheet.NeedsCommit;
-                            if (psheet.UnfocusSelection() && fwdReturn && psheet.SelectedGridEntry != null)
-                            {
-                                psheet.SelectedGridEntry.OnValueReturnKey();
-                            }
-                            return true;
-                        case Keys.Escape:
-                            psheet.OnEscape(this);
-                            return true;
-                        case Keys.F4:
-                            psheet.F4Selection(true);
-                            return true;
-                    }
-                }
-
-                // for the tab key, we want to commit before we allow it to be processed.
-                if ((keyData & Keys.KeyCode) == Keys.Tab && ((keyData & (Keys.Control | Keys.Alt)) == 0))
-                {
-                    return !psheet._Commit();
-                }
-
-                return base.ProcessDialogKey(keyData);
-            }
-
-            protected override void SetVisibleCore(bool value)
-            {
-                Debug.WriteLineIf(CompModSwitches.DebugGridView.TraceVerbose, "DropDownHolder:Visible(" + (value.ToString()) + ")");
-                // make sure we dont' have the mouse captured if
-                // we're going invisible
-                if (value == false && HookMouseDown)
-                {
-                    mouseHook.HookMouseDown = false;
-                }
-                base.SetVisibleCore(value);
-            }
-
-            // a mini version of process dialog key
-            // for responding to WM_GETDLGCODE
-            internal bool WantsTab(bool forward)
-            {
-                return psheet.WantsTab(forward);
-            }
-
-            private unsafe bool WmNotify(ref Message m)
-            {
-                if (m.LParam != IntPtr.Zero)
-                {
-                    User32.NMHDR* nmhdr = (User32.NMHDR*)m.LParam;
-
-                    if (nmhdr->hwndFrom == psheet.ToolTip.Handle)
-                    {
-                        switch ((ComCtl32.TTN)nmhdr->code)
-                        {
-                            case ComCtl32.TTN.SHOW:
-                                PropertyGridView.PositionTooltip(this, psheet.ToolTip, ClientRectangle);
-                                m.Result = (IntPtr)1;
-                                return true;
-                            default:
-                                psheet.WndProc(ref m);
-                                break;
-                        }
-                    }
-                }
-                return false;
-            }
-
-            protected override void WndProc(ref Message m)
-            {
-                if (filter)
-                {
-                    if (psheet.FilterEditWndProc(ref m))
-                    {
-                        return;
-                    }
-                }
-
-                switch ((User32.WM)m.Msg)
-                {
-                    case User32.WM.STYLECHANGED:
-                        if ((unchecked((int)(long)m.WParam) & (int)User32.GWL.EXSTYLE) != 0)
-                        {
-                            psheet.Invalidate();
-                        }
-                        break;
-                    case User32.WM.MOUSEMOVE:
-                        if (unchecked((int)(long)m.LParam) == lastMove)
-                        {
-                            return;
-                        }
-                        lastMove = unchecked((int)(long)m.LParam);
-                        break;
-                    case User32.WM.DESTROY:
-                        mouseHook.HookMouseDown = false;
-                        break;
-                    case User32.WM.SHOWWINDOW:
-                        if (IntPtr.Zero == m.WParam)
-                        {
-                            mouseHook.HookMouseDown = false;
-                        }
-                        break;
-                    case User32.WM.PASTE:
-                        if (ReadOnly)
-                        {
-                            return;
-                        }
-                        break;
-
-                    case User32.WM.GETDLGCODE:
-
-                        m.Result = (IntPtr)((long)m.Result | (int)User32.DLGC.WANTARROWS | (int)User32.DLGC.WANTCHARS);
-                        if (psheet.NeedsCommit || WantsTab((ModifierKeys & Keys.Shift) == 0))
-                        {
-                            m.Result = (IntPtr)((long)m.Result | (int)User32.DLGC.WANTALLKEYS | (int)User32.DLGC.WANTTAB);
-                        }
-                        return;
-
-                    case User32.WM.NOTIFY:
-                        if (WmNotify(ref m))
-                        {
-                            return;
-                        }
-
-                        break;
-                }
-                base.WndProc(ref m);
-            }
-
-            public virtual bool InSetText()
-            {
-                return fInSetText;
-            }
-
-            protected class GridViewEditAccessibleObject : ControlAccessibleObject
-            {
-                private readonly PropertyGridView propertyGridView;
-
-                public GridViewEditAccessibleObject(GridViewEdit owner) : base(owner)
-                {
-                    propertyGridView = owner.psheet;
-                }
-
-                public override AccessibleStates State
-                {
-                    get
-                    {
-                        AccessibleStates states = base.State;
-                        if (IsReadOnly)
-                        {
-                            states |= AccessibleStates.ReadOnly;
-                        }
-                        else
-                        {
-                            states &= ~AccessibleStates.ReadOnly;
-                        }
-                        return states;
-                    }
-                }
-
-                internal override bool IsIAccessibleExSupported() => true;
-
-                /// <summary>
-                ///  Returns the element in the specified direction.
-                /// </summary>
-                /// <param name="direction">Indicates the direction in which to navigate.</param>
-                /// <returns>Returns the element in the specified direction.</returns>
-                internal override UiaCore.IRawElementProviderFragment FragmentNavigate(UiaCore.NavigateDirection direction)
-                {
-                    if (direction == UiaCore.NavigateDirection.Parent && propertyGridView.SelectedGridEntry != null)
-                    {
-                        return propertyGridView.SelectedGridEntry.AccessibilityObject;
-                    }
-                    else if (direction == UiaCore.NavigateDirection.NextSibling)
-                    {
-                        if (propertyGridView.DropDownButton.Visible)
-                        {
-                            return propertyGridView.DropDownButton.AccessibilityObject;
-                        }
-                        else if (propertyGridView.DialogButton.Visible)
-                        {
-                            return propertyGridView.DialogButton.AccessibilityObject;
-                        }
-                    }
-
-                    return base.FragmentNavigate(direction);
-                }
-
-                /// <summary>
-                ///  Gets the top level element.
-                /// </summary>
-                internal override UiaCore.IRawElementProviderFragmentRoot FragmentRoot
-                {
-                    get
-                    {
-                        return propertyGridView.AccessibilityObject;
-                    }
-                }
-
-                internal override object GetPropertyValue(UiaCore.UIA propertyID)
-                {
-                    switch (propertyID)
-                    {
-                        case UiaCore.UIA.RuntimeIdPropertyId:
-                            return RuntimeId;
-                        case UiaCore.UIA.ControlTypePropertyId:
-                            return UiaCore.UIA.EditControlTypeId;
-                        case UiaCore.UIA.NamePropertyId:
-                            return Name;
-                        case UiaCore.UIA.HasKeyboardFocusPropertyId:
-                            return Owner.Focused;
-                        case UiaCore.UIA.IsEnabledPropertyId:
-                            return !IsReadOnly;
-                        case UiaCore.UIA.ClassNamePropertyId:
-                            return Owner.GetType().ToString();
-                        case UiaCore.UIA.FrameworkIdPropertyId:
-                            return NativeMethods.WinFormFrameworkId;
-                        case UiaCore.UIA.IsValuePatternAvailablePropertyId:
-                            return IsPatternSupported(UiaCore.UIA.ValuePatternId);
-                    }
-
-                    return base.GetPropertyValue(propertyID);
-                }
-
-                internal override bool IsPatternSupported(UiaCore.UIA patternId)
-                {
-                    if (patternId == UiaCore.UIA.ValuePatternId)
-                    {
-                        return true;
-                    }
-
-                    return base.IsPatternSupported(patternId);
-                }
-
-                internal override UiaCore.IRawElementProviderSimple HostRawElementProvider
-                {
-                    get
-                    {
-                        // Prevent sending same runtime ID for all edit boxes. Individual edit in
-                        // each row should have unique runtime ID to prevent incorrect announcement.
-                        // For instance screen reader may announce row 2 for the third row edit
-                        // as the sme TextBox control is used both in row 2 and row 3.
-                        return null;
-                    }
-                }
-
-                public override string Name
-                {
-                    get
-                    {
-                        string name = Owner.AccessibleName;
-                        if (name != null)
-                        {
-                            return name;
-                        }
-                        else
-                        {
-                            GridEntry selectedGridEntry = propertyGridView.SelectedGridEntry;
-                            if (selectedGridEntry != null)
-                            {
-                                return selectedGridEntry.AccessibilityObject.Name;
-                            }
-                        }
-
-                        return base.Name;
-                    }
-                    set => base.Name = value;
-                }
-
-                internal override int[] RuntimeId
-                {
-                    get
-                    {
-                        var selectedGridEntryAccessibleRuntimeId =
-                            propertyGridView?.SelectedGridEntry?.AccessibilityObject?.RuntimeId;
-
-                        if (selectedGridEntryAccessibleRuntimeId == null)
-                        {
-                            return null;
-                        }
-
-                        int[] runtimeId = new int[selectedGridEntryAccessibleRuntimeId.Length + 1];
-                        for (int i = 0; i < selectedGridEntryAccessibleRuntimeId.Length; i++)
-                        {
-                            runtimeId[i] = selectedGridEntryAccessibleRuntimeId[i];
-                        }
-
-                        runtimeId[runtimeId.Length - 1] = 1;
-
-                        return runtimeId;
-                    }
-                }
-
-                #region IValueProvider
-
-                internal override bool IsReadOnly
-                {
-                    get
-                    {
-                        return !(propertyGridView.SelectedGridEntry is PropertyDescriptorGridEntry propertyDescriptorGridEntry) || propertyDescriptorGridEntry.IsPropertyReadOnly;
-                    }
-                }
-
-                #endregion
-
-                internal override void SetFocus()
-                {
-                    RaiseAutomationEvent(UiaCore.UIA.AutomationFocusChangedEventId);
-
-                    base.SetFocus();
-                }
-            }
-        }
-
         internal interface IMouseHookClient
         {
             // return true if the click is handled, false
